+++ src/pac/mod.rs
@@ -7,6 +7,7 @@
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 #![allow(unused)]
+#![allow(clippy::upper_case_acronyms)]
 use core::marker::PhantomData;
 use core::ops::Deref;